requirements.txt
pymorhpy2
pymorphy2-dicts-uk